% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filesave.R
\name{formatFiletype}
\alias{formatFiletype}
\title{Formats the value of the filetype argument}
\usage{
formatFiletype(filetype)
}
\arguments{
\item{filetype}{A named list of file extensions or NULL or NA}
}
\value{
A string describing the input value in json format
}
\description{
This function is intended to format the filetype argument of
\code{\link{shinySaveButton}} into a json string representation, so that it
can be attached to the button.
}

