% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shinyFileChoose}
\alias{shinyFileChoose}
\title{Creates a reactive expression that updates the filesystem view}
\usage{
shinyFileChoose(input, inputId, ...)
}
\arguments{
\item{input}{The input object of the \code{shinyServer()} call (usaully
\code{input})}

\item{inputId}{The same ID as used in the matching call to
\code{shinyFilesButton} or as the id attribute of the button, in case of a
manually defined html.}

\item{root}{An absolute filepath giving the root of the returned function}

\item{restrictions}{A vector of directories within the root that should be
filtered out of the results}

\item{filetypes}{A character vector of file extensions (without dot in front
i.e. 'txt' not '.txt') to include in the output. Use the empty string to
include files with no extension. If not set all file types will be included}

\item{hidden}{A logical value specifying whether hidden files should be
returned or not}
}
\value{
A reactive expression that should be assigned to the output object of
the \code{shinyServer()} call. The output name it should be assigned to have
to match the provided \code{inputId}
}
\description{
This function sets up the required connection to the client in order for the
user to navigate the filesystem. For this to work a matching button should be
present in the html, either by using \code{shinyFilesButton()} or adding it
manually. See \code{\link{shinyFilesButton}} for more information on this.
}
\details{
Restrictions on the access rights of the client can be given in several ways.
The root parameter specifies the starting position for the filesystem as
presented to the client. This means that the client can only navigate in
subdirectories of the root. Paths passed of to the \code{restrictions}
parameter will not show up in the client view, and it is impossible to
navigate into these subdirectories. The \code{filetypes} parameter takes a
vector of of file extensions to filter the output on, so that the client is
only presented with these filetypes. The \code{hidden} parameter toggles
whether hidden files should be visible or not.
}
\examples{
\dontrun{
ui <- shinyUI(bootstrapPage(
    shinyFilesButton('files', 'File select', 'Please select a file', FALSE)
))
server <- shinyServer(function(input, output) {
    output$files <- shinyFileChoose(input, 'files', root='.', filetypes=c('', '.txt'))
})

runApp(list(
    ui=ui,
    server=server
))
}
}
\seealso{
Other shinyFiles: \code{\link{parseFilePaths}};
  \code{\link{shinyFilesButton}};
  \code{\link{shinyFilesExample}}
}

