% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFiles-package.R
\docType{package}
\name{shinyFiles-package}
\alias{shinyFiles}
\alias{shinyFiles-package}
\title{A Server-Side File System Viewer for Shiny}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides functionality for client-side navigation of
the server side file system in shiny apps. In case the app is running
locally this gives the user direct access to the file system without the
need to "download" files to a temporary location. Both file and folder
selection as well as file saving is available.
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

Authors:
\itemize{
  \item Vincent Nijs
}

}
