% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{shinyFilesExample}
\alias{shinyFilesExample}
\title{Run a simple example app using the shinyFiles functionality}
\usage{
shinyFilesExample()
}
\description{
When the function is invoked a shiny app is started showing a very simple
setup using shinyFiles. A button summons the dialog box allowing the user to
navigate the R installation directory. To showcase the restrictions parameter
the base package location has been hidden, and is thus inaccessible. A panel
besides the button shows how the user selection is made accessible to the
server after parsing with \code{\link[=parseFilePaths]{parseFilePaths()}}.
}
\seealso{
Other shinyFiles: 
\code{\link{shinyFiles-buttons}},
\code{\link{shinyFiles-observers}},
\code{\link{shinyFiles-parsers}}
}
\concept{shinyFiles}
