% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirchoose.R
\name{updateChildren}
\alias{updateChildren}
\title{Update the children element to reflect current state}
\usage{
updateChildren(oldChildren, currentChildren)
}
\arguments{
\item{oldChildren}{A list of children folders from the parent$children
element of tree in \code{\link[=traverseDirs]{traverseDirs()}}}

\item{currentChildren}{A vector of names of the folders that are currently
present in the parent of oldChildren}
}
\value{
An updated list equal in format to oldChildren
}
\description{
This function create new entries for new folders and remove entries for no
longer existing folders, while keeping the state of transient folders in the
children element of the tree structure. The function does not recurse into
the folders, but merely creates a shell that traverseDirs can take as input.
}
