% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFiles-package.R
\docType{package}
\name{shinyFiles-package}
\alias{shinyFiles}
\alias{shinyFiles-package}
\title{A Server-Side File System Viewer for Shiny}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides functionality for client-side navigation of the server side file system in shiny apps. In case the app is running locally this gives the user direct access to the file system without the need to "download" files to a temporary location. Both file and folder selection as well as file saving is available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/shinyFiles}
  \item Report bugs at \url{https://github.com/thomasp85/shinyFiles/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Vincent Nijs
  \item Thomas Schaffner
  \item Eric Nantz
}

}
