% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filechoose.R
\name{shinyFilesExample}
\alias{shinyFilesExample}
\title{Run a simple example app using the shinyFiles functionality}
\usage{
shinyFilesExample()
}
\description{
When the function is invoked a shiny app is started showing a very simple
setup using shinyFiles. A button summons the dialog box allowing the user to
navigate the R installation directory. To showcase the restrictions parameter
the base package location has been hidden, and is thus inaccecible. A panel
besides the button shows how the user selection is made accessible to the
server after parsing with \code{\link{parseFilePaths}}.
}
\seealso{
Other shinyFiles: \code{\link{parseFilePaths}};
  \code{\link{shinyFileChoose}};
  \code{\link{shinyFilesButton}}
}

